feat(editor): add component favourites, keyboard shortcuts, and netlist preview panel
