+++ eda-frontend/src/components/SchematicEditor/Helper/ComponentDrag.js
@@ -88,6 +88,7 @@ export default function LoadGrid (container, sidebar, outline) {
 
     // Creates the graph inside the given container
     graph = new mxGraph(container)
+    container.graph = graph // Expose for NetlistPreviewPanel
    
 
     mxConnectionHandler.prototype.movePreviewAway = false
