+++ eda-frontend/src/pages/SchematiEditor.js
@@ -11,6 +11,7 @@ import LayoutMain from '../components/Shared/LayoutMain'
 import SchematicToolbar from '../components/SchematicEditor/SchematicToolbar'
 import RightSidebar from '../components/SchematicEditor/RightSidebar'
 import PropertiesSidebar from '../components/SchematicEditor/PropertiesSidebar'
+import NetlistPreviewPanel from '../components/SchematicEditor/NetlistPreviewPanel'
 import LoadGrid from '../components/SchematicEditor/Helper/ComponentDrag.js'
 import ComponentProperties from '../components/SchematicEditor/ComponentProperties'
 import '../components/SchematicEditor/Helper/SchematicEditor.css'
@@ -95,6 +96,7 @@ export default function SchematiEditor (props) {
       {/* Schematic editor Right side pane */}
       <RightSidebar mobileOpen={mobileOpen} mobileClose={handleDrawerToggle}>
         <PropertiesSidebar gridRef={gridRef} outlineRef={outlineRef} />
+        <NetlistPreviewPanel gridRef={gridRef} />
       </RightSidebar>
       <ComponentProperties/>
     </div>
